requirements.txt
pip~=25.1.1
attrs~=25.3.0
wheel~=0.41.2
protobuf~=4.23.4
cryptography~=45.0.3
redis~=6.2.0
setuptools~=80.9.0
packaging~=24.2
pyparsing~=3.2.3
Jinja2~=3.1.6
typing_extensions~=4.14.0
matplotlib~=3.10.3
confluent-kafka~=2.10.0
pandas~=2.2.3
plotly~=6.1.2
streamlit~=1.34.0
pillow~=9.1.1
python-dateutil~=2.9.0.post0
python-dotenv~=1.1.0
pm4py~=2.7.15.2
apache-flink~=1.19.0
river~=0.22.0
logaut~=0.2.0
pylogics~=0.2.1
pythomata~=0.3.2
numpy~=1.24.4
streamlit_js_eval~=0.1.7
httpx~=0.28.1
authlib~=1.6.0